test: add GrpcPlainRequestLoggingInterceptor and add it to BackendResources (#2534)
Grpc's built-in logging is very robust, but does not print request/response messages in "plain text", instead logging byte codes sequences. This interceptor instead logs the protobuf messages in their normal plain text rendering. By default, the logging is disabled. The category needs to be set to CONFIG or higher to see the logs.