fix: update grpc bidi resumable uploads to validate ack'd object size (#2570)
* fix: update grpc bidi resumable uploads to validate ack'd object size

Follow up to #2527

Part 1, add tests and simulation server

* fix: update grpc bidi resumable uploads to validate ack'd object size

Follow up to #2527

Cleanup unused inner-class

* fix: update grpc bidi resumable uploads to validate ack'd object size

Follow up to #2527

Move request trimming inline rather than at the end

* fix: update grpc bidi resumable uploads to validate ack'd object size

Follow up to #2527

Update GapicBidiUnbufferedWritableByteChannel to correctly handle failure scenarios {1, 2, 3, 4, 4_1, 4_2, 5, 7}

Refactor ResumableSessionFailureScenario to accept Message and List rather than WriteObjectResponse and List<WriteObjectRequest>.

The shape of BidiWriteObjectRequest/WriteObjectRequest and BidiWriteObjectResponse/WriteObjectResponse are largely the same (bidi has two extra fields) so rather than overloading toStorageException yet again, this time there is a single method for grpc messages that internally can branch when formatting the request. In this case, we're quite safe taking this relaxed typing because it is an internal API that is only called from a grpc context where protos will be used.

* fix: update grpc bidi resumable uploads to validate ack'd object size

Follow up to #2527

Update GapicBidiUnbufferedWritableByteChannel to correctly handle partial consumption of content

Tests passing now.

Separate tracking of client detected errors and stream errors.

When await is invoked, if a client detected error is present AND a stream error is present, the client detected error will take precedence. If a stream error happens and the client detected error has not yet been observed, the stream error will be added as a suppressed exception to the client detected error.

* chore: fix failing tests