chore: ensure generation is included when a grpc ReadObject is retried (#2741)
If a read is opened without a generation, we want to ensure we restart for the same generation. We have this behavior for json

This is mainly to avoid a race condition of reading an object (gen=1), that object being replaced (gen=2) then retry attempts to read it again we need to read gen=1 otherwise we will return corrupted data.