chore: allow a fast opened ObjectReadSession to be used for additional reads
Currently, a FastOpenObjectReadSession will be closed when it's resource is closed. This change will allow for the underlying ObjectReadSession to be held open even when the initial projection is closed.

This allows for the same stream to be used in the case when a "seek" might need to be performed.