chore: refactor RangeProjectionConfigs to have RangeSpec provided as part of the config (#99)
Not all projections have the same way of determining a RangeSpec for its operation. As such, RangeSpec needs to be provided per projection rather than passed in to the BlobReadSession readAs method. For example, a seekable channel sets it's offset by seeking, whereas a range read as an ApiFuture<byte[]> will want a RangeSpec.