chore: fixup elapsed reporting in retry related messages
Update Backoff so that it doesn't assume the returned backoff is fully consumed, instead expect that to be included in the provided `elapsed` value if/when called again. Update backoff tests to pass the previous backoff result into the next call.

Update DefaultRetryContext to track elapsed time since last recordError invocation separately from the last time the context was reset.

Update RetryContextTest to have new assertions for the fixed elapsed durations.