build: move hermetic_library_generation filtering logic into the steps rather than on.pull_request
When filtering with on.pull_request the workflow will not report a status if it does not match. This causes contention with this workflow being declared as a required status. To sidestep this, move the matching logic into the steps of the workflow, similar to how we filter out PRs from forks.